renamed test utilities and made them private
