+++ doubleml/tests/_utils_iivm_manual.py
@@ -2,7 +2,7 @@ import numpy as np
 from sklearn.model_selection import KFold, GridSearchCV
 from sklearn.base import clone
 
-from doubleml.tests.helper_boot import boot_manual, draw_weights
+from ._utils_boot import boot_manual, draw_weights
 
 
 def fit_nuisance_iivm(y, x, d, z, learner_m, learner_g, learner_r, smpls,
