+++ doubleml/tests/test_plr_no_cross_fit.py
@@ -9,7 +9,7 @@ from sklearn.linear_model import Lasso
 
 import doubleml as dml
 
-from doubleml.tests.helper_plr_manual import plr_dml1, fit_nuisance_plr, boot_plr, tune_nuisance_plr
+from ._utils_plr_manual import plr_dml1, fit_nuisance_plr, boot_plr, tune_nuisance_plr
 
 
 @pytest.fixture(scope='module',
