+++ doubleml/tests/test_plr_rep_cross.py
@@ -10,7 +10,7 @@ from sklearn.ensemble import RandomForestRegressor
 
 import doubleml as dml
 
-from doubleml.tests.helper_plr_manual import plr_dml1, plr_dml2, fit_nuisance_plr, boot_plr
+from ._utils_plr_manual import plr_dml1, plr_dml2, fit_nuisance_plr, boot_plr
 
 
 @pytest.fixture(scope='module',
